Split support classes/traits from MultiBackendTest.
