+++ module/VuFind/tests/unit-tests/src/VuFindTest/ILS/Driver/MultiBackendTest/DemoMock.php
@@ -1,0 +1,7 @@
+<?php
+namespace VuFindTest\ILS\Driver;
+
+class DemoMock extends \VuFind\ILS\Driver\Demo
+{
+    use ILSMockTrait;
+}
